Tunnel: Automatically marshall record types (#2)
Adds support for automatically marshalling `record` types that have marshallable record components. Also allows specifying `enum` and `record` types in the `MarshallAs` annotation (required for lists), and an `Identifier` marshaller (built-in to `PacketByteBuf` and used in my tests).

Annotations specified on record components are also applied to their fields, and therefore work here, including `@Optional` and `@MarshalledAs`. Records may, directly or indirectly, contain instances of themselves as components, but may not (in)directly contain *references* to themselves.

`@MarshallAs` may refer to the fully qualified name of a `record` or `enum` type, allowing you to serialize lists of enum or record values.

Note that record fields are *not* bit-packed, to simplify the implementation.

Also bumps version to 0.2.2.

Co-authored-by: l-Luna <izzeldeen03@gmail.com>
Reviewed-on: https://git.sleeping.town/unascribed/Lib39/pulls/2
Co-authored-by: l-Luna <l-luna@noreply.git.sleeping.town>
Co-committed-by: l-Luna <l-luna@noreply.git.sleeping.town>